\name{mi-internal}
\alias{mi-internal}
\alias{change,missing_data.frame,ANY,ANY,character-method}
\alias{mi,missing_data.frame,missing-method}
\alias{plot,missing_data.frame,missing-method}
\alias{plot,mi,ANY-method}
\alias{show,missing_data.frame-method}
\alias{show,missing_variable-method}
\alias{summary,missing_data.frame-method}
\alias{summary,mi-method}
\alias{show,mi-method}

\alias{change_family,missing,missing_variable,family-method}
\alias{change_family,missing,proportion,family-method}
\alias{change_family,missing,unordered-categorical,family-method}
\alias{change_imputation_method,missing,missing_variable,character-method}
\alias{change_imputation_method,missing,missing_variable,missing-method}
\alias{change_link,missing,missing_variable,character-method}
\alias{change_link,missing,missing_variable,missing-method}
\alias{change_link,missing_data.frame,character,character-method}
\alias{change_link,missing_data.frame,numeric,character-method}
\alias{change_link,missing_data.frame,logical,character-method}
\alias{change_model,missing,missing_variable,character-method}
\alias{change_model,missing_data.frame,character,character-method}
\alias{change_model,missing_data.frame,numeric,character-method}
\alias{change_model,missing_data.frame,logical,character-method}
\alias{change_size,missing,missing_variable,integer-method}
\alias{change_size,missing,categorical,integer-method}
\alias{change_size,missing,fixed,integer-method}
\alias{change_size,missing_data.frame,missing,integer-method}
\alias{change_type,missing,missing_variable,character-method}
\alias{change_imputation_method,missing_data.frame,logical,character-method}
\alias{change_imputation_method,missing_data.frame,numeric,character-method}
\alias{change_transformation,missing,missing_variable,missing-method}
\alias{change_transformation,missing_data.frame,character,missing-method}
\alias{change_transformation,missing_data.frame,character,function-method}
\alias{change_transformation,missing_data.frame,numeric,function-method}
\alias{change_transformation,missing_data.frame,logical,function-method}
\alias{change_type,missing,missing_variable,missing-method}
\alias{change_type,missing_data.frame,character,missing-method}
\alias{complete,irrelevant,integer-method}
\alias{complete,categorical,integer-method}
\alias{complete,binary,integer-method}
\alias{complete,continuous,integer-method}
\alias{complete,nonnegative-continuous,integer-method}
\alias{complete,SC_proportion,integer-method}
\alias{complete,mi,integer-method}
\alias{complete,missing_data.frame,numeric-method}
\alias{complete,missing_variable,integer-method}
\alias{complete,mi,missing-method}
\alias{complete,mi_list,numeric-method}
\alias{complete,mi_list,missing-method}
\alias{fit_model,missing_variable,missing_data.frame-method}
\alias{fit_model,missing,missing_data.frame-method}
\alias{fit_model,continuous,missing_data.frame-method}
\alias{fit_model,semi-continuous,missing_data.frame-method}
\alias{fit_model,nonnegative-continuous,missing_data.frame-method}
\alias{fit_model,SC_proportion,missing_data.frame-method}
\alias{fit_model,proportion,missing_data.frame-method}
\alias{fit_model,truncated-continuous,missing_data.frame-method}
\alias{fit_model,censored-continuous,missing_data.frame-method}
\alias{fit_model,missing_variable,experiment_missing_data.frame-method}
\alias{fit_model,continuous,experiment_missing_data.frame-method}
\alias{fit_model,missing,multilevel_missing_data.frame-method}
\alias{fit_model,missing,mdf_list-method}
\alias{fit_model,binary,allcategorical_missing_data.frame-method}
\alias{fit_model,missing,allcategorical_missing_data.frame-method}
\alias{fit_model,ordered-categorical,allcategorical_missing_data.frame-method}
\alias{fit_model,unordered-categorical,allcategorical_missing_data.frame-method}
\alias{get_parameters,ANY-method}
\alias{get_parameters,polr-method}
\alias{get_parameters,multinom-method}
\alias{get_parameters,mi-method}
\alias{get_parameters,mi_list-method}
\alias{get_parameters,missing_data.frame-method}
\alias{get_parameters,missing_variable-method}
\alias{hist,mi-method}
\alias{hist,missing_variable-method}
\alias{hist,semi-continuous-method}
\alias{hist,binary-method}
\alias{hist,categorical-method}
\alias{initialize,missing_variable-method}
\alias{image,mi-method}
\alias{image,mi_list-method}
\alias{image,mdf_list-method}
\alias{image,missing_data.frame-method}
\alias{image,dgTMatrix-method}
\alias{mi,character,missing-method}
\alias{mi,missing_variable,ANY-method}
\alias{mi,missing_variable,missing-method}
\alias{mi,semi-continuous,missing-method}
\alias{mi,bounded-continuous,missing-method}
\alias{mi,binary,glm-method}
\alias{mi,grouped-binary,clogit-method}
\alias{mi,continuous,glm-method}
\alias{mi,bounded-continuous,glm-method}
\alias{mi,SC_proportion,betareg-method}
\alias{mi,proportion,betareg-method}
\alias{mi,nonnegative-continuous,glm-method}
\alias{mi,censored-continuous,glm-method}
\alias{mi,semi-continuous,glm-method}
\alias{mi,categorical,missing-method}
\alias{mi,count,glm-method}
\alias{mi,data.frame,missing-method}
\alias{mi,irrelevant,ANY-method}
\alias{mi,interval,glm-method}
\alias{mi,mdf_list,missing-method}
\alias{mi,mi_list,missing-method}
\alias{mi,list,missing-method}
\alias{mi,missing_data.frame,mi-method}
\alias{mi,matrix,missing-method}
\alias{mi,mi,missing-method}
\alias{mi,by,missing-method}
\alias{mi,missing_variable,ANY-method}
\alias{mi,missing_variable,missing-method}
\alias{mi,nonnegative-continuous,missing-method}
\alias{mi,ordered-categorical,polr-method}
\alias{mi,proportion,glm-method}
\alias{mi,unordered-categorical,multinom-method}
\alias{mi,unordered-categorical,RNL-method}
\alias{mi,categorical,matrix-method}
\alias{missing_data.frame,data.frame-method}
\alias{missing_data.frame,list-method}
\alias{missing_data.frame,matrix-method}
\alias{missing_variable,ANY,character-method}
\alias{missing_variable,ANY,missing-method}
\alias{betareg-class}
\alias{clogit-class}
\alias{mdf_list-class}
\alias{mi_list-class}
\alias{family-class}
\alias{multinom-class}
\alias{RNL-class}
\alias{plot,missing_data.frame,missing_variable-method}
\alias{plot,mi_list,ANY-method}
\alias{plot,allcategorical_missing_data.frame,binary-method}
\alias{plot,allcategorical_missing_data.frame,categorical-method}
\alias{change_family,missing_data.frame,character,character-method}
\alias{change_family,missing,missing_variable,missing-method}
\alias{change_family,missing_data.frame,character,family-method}
\alias{change_family,missing_data.frame,character,list-method}
\alias{change_family,missing_data.frame,logical,character-method}
\alias{change_family,missing_data.frame,logical,family-method}
\alias{change_family,missing_data.frame,numeric,character-method}
\alias{change_family,missing_data.frame,numeric,family-method}
\alias{change_family,missing_data.frame,numeric,list-method}
\alias{change_imputation_method,missing_data.frame,character,character-method}
\alias{change_size,missing_data.frame,numeric-method}
\alias{change_transformation,missing,missing_variable,function-method}
\alias{change_transformation,missing_data.frame,character,character-method}
\alias{change_transformation,missing_data.frame,numeric,character-method}
\alias{change_transformation,missing_data.frame,logical,character-method}
\alias{change_type,missing_data.frame,character,character-method}
\alias{change_type,missing_data.frame,logical,character-method}
\alias{change_type,missing_data.frame,numeric,character-method}
\alias{change,missing_data.frame,ANY,numeric,numeric-method}
\alias{change,missing_data.frame,ANY,logical,numeric-method}
\alias{change,missing_data.frame,ANY,character,numeric-method}
\alias{change,missing_data.frame,ANY,logical,character-method}
\alias{change,mdf_list,ANY,ANY,ANY-method}
\alias{coerce,data.frame,missing_data.frame-method}
\alias{coerce,matrix,missing_data.frame-method}
\alias{coerce,missing_data.frame,data.frame-method}
\alias{coerce,missing_data.frame,matrix-method}
\alias{complete,missing_data.frame,integer-method}
\alias{complete,mi,numeric-method}
\alias{fit_model,binary,missing_data.frame-method}
\alias{fit_model,grouped-binary,missing_data.frame-method}
\alias{fit_model,count,missing_data.frame-method}
\alias{fit_model,irrelevant,missing_data.frame-method}
\alias{fit_model,interval,missing_data.frame-method}
\alias{fit_model,missing_variable,missing_data.frame-method}
\alias{fit_model,ordered-categorical,missing_data.frame-method}
\alias{fit_model,unordered-categorical,missing_data.frame-method}
\alias{fit_model,character,mi-method}
\alias{fit_model,missing,mi-method}
\alias{fit_model,missing_data.frame,missing_data.frame-method}
\alias{hist,missing_data.frame-method}
\alias{hist,mdf_list-method}
\alias{hist,mi_list-method}
\alias{initialize,missing_data.frame-method}
\alias{plot,missing_data.frame,binary-method}
\alias{plot,missing_data.frame,categorical-method}
\alias{plot,missing_data.frame,semi-continuous-method}
\alias{plot,missing_data.frame,missing_variable-method}
\alias{plot,mi,missing-method}
\alias{traceplot,mi}
\alias{traceplot,mi_list}
\alias{.prune}
\alias{.possible_missing_variable}
\title{Internal Functions and Methods}
\description{
These functions are not intended to be called directly. In the case of
methods, they documented elsewhere, either with the associated generic
function or with the class of the object that the method is defined for.
}
\keyword{internal}