\name{type.models}
\alias{type.models}
\alias{mi.types}

\title{Functions to identify types of the models of the mi object}
\description{
The function to select the model based on the variable type
}
\usage{
type.models(type)
mi.types()

}

\arguments{
  \item{type}{Nine types: \code{continuous}, 
        \code{ordered-categorical}, 
        \code{unordered-categorical}, 
        \code{dichotomous},
        \code{mixed},
        \code{redictive-mean-match},
        \code{fixed}}
}

\author{
  Masanao Yajima \email{yajima@stat.columbia.edu},
  Yu-Sung Su \email{ys463@columbia.edu},
  Andrew Gelman \email{gelman@stat.columbia.edu}
}


\seealso{\code{\link{mi.info}}}

\keyword{manip}
